Refactor JOSE headers for better handling
